--- src/components/OnAddComponent/index.jsx
@@ -1,40 +1,0 @@
-import { useState } from "react";
-import Basket from "components/Basket";
-
-const OnAddComponent = (props) => {
-  const { products } = props
-  
-  const { onAdd, onRemove, cartItems } = props
-
-  // const [cartItems, setCartItems] = useState([]);
-
-  // const onAdd = (product) => {
-  //   const exist = cartItems.find((x) => x.id === product.id);
-  //   if (exist) {
-  //     setCartItems(
-  //       cartItems.map((x) =>
-  //         x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
-  //       )
-  //     );
-  //   } else {
-  //     setCartItems([...cartItems, { ...product, qty: 1 }]);
-  //   }
-  // };
-  // const onRemove = (product) => {
-  //   const exist = cartItems.find((x) => x.id === product.id);
-  //   if (exist.qty === 1) {
-  //     setCartItems(cartItems.filter((x) => x.id !== product.id));
-  //   } else {
-  //     setCartItems(
-  //       cartItems.map((x) =>
-  //         x.id === product.id ? { ...exist, qty: exist.qty - 1 } : x
-  //       )
-  //     );
-  //   }
-  // };
-  return (
-    <Basket cartItems={cartItems} onAdd={onAdd} onRemove={onRemove}></Basket>
-  );
-};
-
-export default OnAddComponent;
