+++ src/components/Product/index.jsx
@@ -1,5 +1,7 @@
-import React from 'react';
+import React, {useContext} from 'react';
 import { Link } from 'react-router-dom';
+import { CartContext } from "../../contexts/cart";
+import data from 'mocks/products/items';
 
 import {
   AddProductCart, Card, CardBtn, CardImage, CardLeft, CardRating, CardRight, HeartPlus, StarRating, Text
@@ -7,7 +9,9 @@ import {
 
 
 export default function Product(props) {
-  const { product, onAdd } = props;
+  const { product } = props;
+  const { products } = data;
+  const { cartItems, onAdd, onRemove } = useContext(CartContext);
   // const { productCategory, productName, productId } = useParams()
 
   return (
