+++ src/pages/Shop/index.jsx
@@ -1,55 +1,27 @@
-import React, { useState } from "react";
-import Basket from "components/Basket";
+import React, { useState, useContext } from "react";
 import ProductsList from "components/ProductsList";
 import data from "../../mocks/products/items";
-import OnAddComponent from "components/OnAddComponent";
+import { CartContext } from "../../contexts/cart";
+import Basket from "components/Basket";
 
 export const Shop = () => {
   const { products } = data;
-  
-  // const { onAdd, onRemove, cartItems } = props
 
-  const [cartItems, setCartItems] = useState([]);
+  const { cartItems, onAdd, onRemove } = useContext(CartContext);
 
-  const onAdd = (product) => {
-    const exist = cartItems.find((x) => x.id === product.id);
-    if (exist) {
-      setCartItems(
-        cartItems.map((x) =>
-          x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
-        )
-      );
-    } else {
-      setCartItems([...cartItems, { ...product, qty: 1 }]);
-    }
-  };
-  const onRemove = (product) => {
-    const exist = cartItems.find((x) => x.id === product.id);
-    if (exist.qty === 1) {
-      setCartItems(cartItems.filter((x) => x.id !== product.id));
-    } else {
-      setCartItems(
-        cartItems.map((x) =>
-          x.id === product.id ? { ...exist, qty: exist.qty - 1 } : x
-        )
-      );
-    }
-  };
   return (
     <>
-      <ProductsList products={products} cartItems={cartItems}
-        onAdd={onAdd}
-        onRemove={onRemove}/>
-
-      <OnAddComponent
+      <ProductsList
+        products={products}
         cartItems={cartItems}
         onAdd={onAdd}
         onRemove={onRemove}
-      ></OnAddComponent>
+      />
 
-    </>
+      <Basket cartItems={cartItems} onAdd={onAdd} onRemove={onRemove}></Basket>
+    
+      </>
   );
+};
 
-}
-
-export default Shop
+export default Shop;
